Separate generater into new class
