+++ lib/ffi_gen/generator/java_pre.java
